Added comments, refactored folder structure, small changes to dashboard navigation
