+++ src/components/dashboard/Dashboard.js
@@ -1,19 +1,29 @@
+// React Components
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import NavBar from './dashboard/NavBar';
-import { getTickets } from '../store/actions/';
-import Tickets from './Tickets';
-import UserModal from './UserModal'
+// Store
+import { getTickets } from '../../store/actions';
+
+// Components
+import NavBar from './Navigation/NavBar';
+import Tickets from './Tickets/Tickets';
+import UserModal from './Users/UserModal';
+
 class App extends Component {
+  // Sets state for Dashboard component
   state = {
     show: false
-  }
+  };
+
+  // ShowModal function
   showModal = evt => {
     this.setState({
       show: !this.state.show
-    })
-  }
+    });
+  };
+
+  //CDM, calls getTickets function from Redux
   componentDidMount() {
     this.props.getTickets();
   }
@@ -21,7 +31,10 @@ class App extends Component {
   render() {
     return (
       <>
+        {/* Renders NavBar component */}
         <NavBar />
+
+        {/* Renders Dashboard then loads tickets using stored state loading */}
         <div className='dashboard'>
           {this.props.loading ? (
             <p>Loading Tickets...</p>
@@ -30,15 +43,21 @@ class App extends Component {
           )}
         </div>
 
-        <button onClick={evt => {
-          this.showModal(evt)
-        }}>{''}View All User{''}</button>
+        {/* modal button to show all users */}
+        <button
+          onClick={evt => {
+            this.showModal(evt);
+          }}
+        >
+          {''}View All User{''}
+        </button>
         <UserModal onClose={this.showModal} show={this.state.show} />
       </>
     );
   }
 }
 
+// From the Redux, we're using the states loading, error, and tickets.
 const mapStateToProps = state => {
   console.log('DASHBOARD STATE:', state);
   return {
@@ -48,6 +67,7 @@ const mapStateToProps = state => {
   };
 };
 
+// From Redux, we're pulling the getTickets function
 const mapDispatchToProps = {
   getTickets
 };
