+++ src/components/dashboard/Tickets/Tickets.js
@@ -1,11 +1,13 @@
+// React Components
 import React, { Component } from 'react';
+
+// Components
 import Ticket from './Ticket';
 
 export default class Tickets extends Component {
   render() {
-    console.log('TICKETS PROPS:', this.props);
-    console.log('TICKETS STATE:', this.state);
     return (
+      // Maps and passes All Tickets to the Ticket Component
       <div className='tickets'>
         <h1 className='tickets-title'>All Tickets</h1>
         {this.props.tickets.map(ticket => {
