+++ src/index.js
@@ -1,11 +1,15 @@
+// React Components
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 
+// Store
+import store from './store/store';
+
+// Components
 import App from './App';
 import './scss/index.scss';
-import store from './store/store';
 
 ReactDOM.render(
   <Provider store={store}>
