+++ src/scss/nav.scss
@@ -1,6 +1,7 @@
 .navbar {
     display: flex;
     align-items: center;
+    justify-content: space-between;
     top: 0;
     height: 70px;
     width: 100%;
@@ -21,11 +22,16 @@
         font-size: 5rem;
         text-decoration: none;
         color: #333;
+        border-left: 3px solid #333;
     }
 
     .wb-user {
         font-size: 4rem;
         font-weight: bold;
         color: #007fb9;
+    }
+
+    .icons-container {
+        display: flex;
     }
 }
